lib: add reject_hidden Hash extension
This is intended to work with some sort of json output filter to enable
omnibus removal of all hidden: true objects from the api in cases where
the viewing user does not have :update? permissions on the resource. I
don't know how to apply this as an output filter, but this at least
captures the idea of what I want to do.

There's tests in spec/lib/, which should explain what I'm trying to
accomplish better than words

This is definitely super WIP and won't work at all with arrays, but it's
a start :)